feat: major deep context quality improvements and repository cleanup
This commit addresses 8 critical quality issues with deep_context.md generation:

• SATD annotations now only appear when technical debt is actually detected
• PageRank calculations are now conditional and connectivity-based instead of hardcoded 0.15
• Overall Health normalized to proper 0-100 TDG scale (was showing 6833%)
• Test Coverage percentages fixed (was showing 6500%)
• Churn annotations properly integrated
• Auto-scaling concurrency implemented for powerful systems
• Shell script AST analysis verified and working
• Comprehensive repository cleanup (48 old files removed)

Technical improvements:
- Fixed PageRank to calculate based on actual graph connectivity
- Implemented conditional SATD detection to reduce noise
- Normalized all percentage displays to meaningful ranges
- Applied EXTREME TDD methodology with comprehensive test coverage
- Achieved ZERO clippy warnings across library and binary
- Fixed 18 failing tests via root cause analysis
- Applied cargo fix for all compilation warnings

Repository maintenance:
- Removed 48 obsolete markdown files and temporary scripts
- Cleaned up root directory structure
- Applied consistent code formatting and style

+++ server/src/tdg/storage_backend.rs
@@ -268,7 +268,7 @@ impl StorageBackend for RocksDbBackend {
         let size = self
             .db
             .property_int_value("rocksdb.live-sst-files-size")?
-            .unwrap_or(0) as u64;
+            .unwrap_or(0);
         Ok(size)
     }
